Refactored Thread.sleep functions. Removed Adyen specific Trip tests as not needed.
All tests pass
