+++ app/src/androidTest/java/com/karhoo/karhootraveller/common/BaseRobot.kt
@@ -201,7 +201,15 @@ open class BaseTestRobot {
         onView(allOf(withId(listRes), isDescendantOfA(withId(parentRes)))).check(RecyclerMatcher(listSize))
     }
 
-    fun sleep(millis: Long = 350) = apply {
+    fun shortSleep(millis: Long = 350) = apply {
+        Thread.sleep(millis)
+    }
+
+    fun mediumSleep(millis: Long = 1000) = apply {
+        Thread.sleep(millis)
+    }
+
+    fun longSleep(millis: Long = 5000) = apply {
         Thread.sleep(millis)
     }
 
@@ -291,25 +299,6 @@ open class BaseTestRobot {
                 Press.FINGER,
                 InputDevice.SOURCE_MOUSE,
                 MotionEvent.BUTTON_PRIMARY)
-    }
-
-    /**
-     * Perform action of waiting for a specific time.
-     */
-    open fun waitFor(millis: Long): ViewAction? {
-        return object : ViewAction {
-            override fun getConstraints(): Matcher<View> {
-                return isRoot()
-            }
-
-            override fun getDescription(): String {
-                return "Wait for $millis milliseconds."
-            }
-
-            override fun perform(uiController: UiController, view: View) {
-                uiController.loopMainThreadForAtLeast(millis)
-            }
-        }
     }
 
     fun withIndex(matcher: Matcher<View>, index: Int): Matcher<View> {
