+++ app/src/androidTest/java/com/karhoo/karhootraveller/menu/MenuTests.kt
@@ -12,8 +12,6 @@ import com.karhoo.karhootraveller.profile.user.userProfile
 import com.karhoo.uisdk.booking.booking
 import com.karhoo.uisdk.common.Launch
 import com.karhoo.uisdk.screen.booking.BookingActivity
-import com.karhoo.uisdk.util.TestData.Companion.LONG
-import com.karhoo.uisdk.util.TestData.Companion.SHORT
 import com.karhoo.uisdk.util.TestData.Companion.USER
 import com.schibsted.spain.barista.rule.flaky.AllowFlaky
 import org.junit.Rule
@@ -59,7 +57,7 @@ class MenuTests : Launch {
         }
         userProfile {
             clickBackToolbarButton()
-            waitFor(SHORT)
+            shortSleep()
         }
         booking {
         } result {
@@ -82,7 +80,7 @@ class MenuTests : Launch {
         }
         menu {
             clickOnHelpButton()
-            sleep(LONG)
+            longSleep()
         } result {
             helpPageIsShown()
         }
@@ -187,7 +185,7 @@ class MenuTests : Launch {
         }
         menu {
             clickOnAboutButton()
-            sleep()
+            shortSleep()
             clickBackToolbarButton()
         }
         booking {
