+++ app/src/androidTest/java/com/karhoo/karhootraveller/profile/user/ProfileLOCTest.kt
@@ -10,7 +10,6 @@ import com.karhoo.karhootraveller.common.networkServiceRobot
 import com.karhoo.karhootraveller.common.preferences
 import com.karhoo.karhootraveller.common.testrunner.TravellerTestConfig
 import com.karhoo.karhootraveller.presentation.profile.ProfileActivity
-import com.karhoo.uisdk.util.TestData.Companion.MEDIUM
 import org.junit.After
 import org.junit.Rule
 import org.junit.Test
@@ -48,7 +47,7 @@ class ProfileLOCTest : Launch {
             networkServiceRobot {
                 disableNetwork(activityRule.activity.applicationContext)
             }
-            sleep(MEDIUM)
+            mediumSleep()
         } result {
             checkSnackbarWithText(R.string.network_error)
         }
