+++ uisdk/src/androidTest/java/com/karhoo/uisdk/booking/GuestBookingTests.kt
@@ -23,8 +23,6 @@ import com.karhoo.uisdk.util.TestData.Companion.BRAINTREE_PROVIDER
 import com.karhoo.uisdk.util.TestData.Companion.BRAINTREE_TOKEN
 import com.karhoo.uisdk.util.TestData.Companion.DESTINATION_TRIP
 import com.karhoo.uisdk.util.TestData.Companion.DRIVER_TRACKING
-import com.karhoo.uisdk.util.TestData.Companion.LONG
-import com.karhoo.uisdk.util.TestData.Companion.MEDIUM
 import com.karhoo.uisdk.util.TestData.Companion.ORIGIN_TRIP
 import com.karhoo.uisdk.util.TestData.Companion.PAYMENTS_TOKEN
 import com.karhoo.uisdk.util.TestData.Companion.PLACE_DETAILS
@@ -33,7 +31,6 @@ import com.karhoo.uisdk.util.TestData.Companion.PLACE_SEARCH_RESULT
 import com.karhoo.uisdk.util.TestData.Companion.PLACE_SEARCH_RESULT_EXTRA
 import com.karhoo.uisdk.util.TestData.Companion.QUOTE_LIST_ID_ASAP
 import com.karhoo.uisdk.util.TestData.Companion.SEARCH_ADDRESS
-import com.karhoo.uisdk.util.TestData.Companion.SHORT
 import com.karhoo.uisdk.util.TestData.Companion.TRIP
 import com.karhoo.uisdk.util.TestData.Companion.TRIP_DER_NO_NUMBER_PLATE
 import com.karhoo.uisdk.util.TestData.Companion.TRIP_STATUS_DER
@@ -108,7 +105,7 @@ class GuestBookingTests : Launch {
             quotesResponse(HTTP_OK, VEHICLES_ASAP)
         }
         booking(this, INITIAL_TRIP_INTENT) {
-            sleep()
+            shortSleep()
         } result {
             fullASAPQuotesListCheckGuest()
         }
@@ -126,7 +123,7 @@ class GuestBookingTests : Launch {
             paymentsProviderResponse(HTTP_OK, BRAINTREE_PROVIDER)
         }
         booking(this) {
-            sleep()
+            shortSleep()
         } result {
             guestCheckoutEmptyFullCheck()
         }
@@ -148,7 +145,7 @@ class GuestBookingTests : Launch {
             clickPickUpAddressField()
         }
         address {
-            sleep()
+            shortSleep()
         } result {
             checkAddressScreenFromPickupGuestCheckout()
         }
@@ -170,7 +167,7 @@ class GuestBookingTests : Launch {
             clickDestinationAddressField()
         }
         address {
-            sleep()
+            shortSleep()
         } result {
             checkAddressScreenFromDestinationGuestCheckout()
         }
@@ -195,7 +192,7 @@ class GuestBookingTests : Launch {
         }
         address {
             search(SEARCH_ADDRESS)
-            waitFor(MEDIUM)
+            mediumSleep()
             clickBakerStreetResult()
         }
         serverRobot {
@@ -207,11 +204,11 @@ class GuestBookingTests : Launch {
         }
         address {
             search(TestData.SEARCH_ADDRESS_EXTRA)
-            waitFor(SHORT)
+            shortSleep()
             clickOxfordStreetResult()
         }
         booking {
-            waitFor(SHORT)
+            shortSleep()
         } result {
             bothSelectedAddressesAreVisible()
         }
@@ -236,7 +233,7 @@ class GuestBookingTests : Launch {
         }
         address {
             search(TestData.SEARCH_ADDRESS)
-            waitFor(SHORT)
+            shortSleep()
             clickBakerStreetResult()
         }
         serverRobot {
@@ -252,11 +249,11 @@ class GuestBookingTests : Launch {
         }
         address {
             search(TestData.SEARCH_ADDRESS_EXTRA)
-            waitFor(SHORT)
+            shortSleep()
             clickOxfordStreetResult()
         }
         booking {
-            waitFor(SHORT)
+            shortSleep()
         } result {
             fullASAPQuotesListCheckGuest()
         }
@@ -298,11 +295,11 @@ class GuestBookingTests : Launch {
         }
         address {
             search(TestData.SEARCH_ADDRESS)
-            sleep()
+            shortSleep()
             clickBakerStreetResult()
         }
         booking {
-            sleep()
+            shortSleep()
         } result {
             flowBookingPickupBookingCheck()
             // TODO check pickupPinIcon
@@ -323,9 +320,9 @@ class GuestBookingTests : Launch {
             quotesResponse(HTTP_OK, VEHICLES_ASAP)
         }
         booking(this, INITIAL_TRIP_INTENT) {
-            sleep()
+            shortSleep()
             pressFirstQuote()
-            sleep()
+            shortSleep()
         } result {
             checkGuestDetailsPageIsShown()
         }
@@ -361,19 +358,19 @@ class GuestBookingTests : Launch {
             guestBookingDetailsResponse(code = HTTP_OK, response = TRIP_DER_NO_NUMBER_PLATE, trip = TRIP.tripId)
         }
         booking(this, INITIAL_TRIP_INTENT) {
-            waitFor(MEDIUM)
+            mediumSleep()
             pressFirstQuote()
-            waitFor(MEDIUM)
+            mediumSleep()
             fillCorrectInfoGuestDetails()
             enterCardDetails()
-            waitFor(LONG)
+            longSleep()
         } result {
             fullCheckFilledGuestDetailsPage()
             guestBookingCheckCardDetails()
         }
         booking {
             pressBookRideButton()
-            waitFor(MEDIUM)
+            mediumSleep()
         } result {
             checkWebViewDisplayed()
         }
