+++ uisdk/src/androidTest/java/com/karhoo/uisdk/booking/adyen/AdyenBookingTests.kt
@@ -13,8 +13,8 @@ import com.karhoo.karhootraveller.profile.user.userProfile
 import com.karhoo.sdk.api.model.TripInfo
 import com.karhoo.uisdk.R
 import com.karhoo.uisdk.address.address
-import com.karhoo.uisdk.booking.braintree.BraintreeBookingTests
 import com.karhoo.uisdk.booking.booking
+import com.karhoo.uisdk.booking.braintree.BraintreeBookingTests
 import com.karhoo.uisdk.common.Launch
 import com.karhoo.uisdk.common.serverRobot
 import com.karhoo.uisdk.common.testrunner.UiSDKTestConfig
@@ -29,8 +29,6 @@ import com.karhoo.uisdk.util.TestData.Companion.DESTINATION_TRIP
 import com.karhoo.uisdk.util.TestData.Companion.DRIVER_TRACKING
 import com.karhoo.uisdk.util.TestData.Companion.FLEET_INFO_ALT
 import com.karhoo.uisdk.util.TestData.Companion.GENERAL_ERROR
-import com.karhoo.uisdk.util.TestData.Companion.LONG
-import com.karhoo.uisdk.util.TestData.Companion.MEDIUM
 import com.karhoo.uisdk.util.TestData.Companion.NO_AVAILABILITY
 import com.karhoo.uisdk.util.TestData.Companion.ORIGIN_TRIP
 import com.karhoo.uisdk.util.TestData.Companion.PLACE_DETAILS
@@ -41,7 +39,6 @@ import com.karhoo.uisdk.util.TestData.Companion.REVERSE_GEO_SUCCESS
 import com.karhoo.uisdk.util.TestData.Companion.REVERSE_GEO_SUCCESS_ALTERNATIVE
 import com.karhoo.uisdk.util.TestData.Companion.SEARCH_ADDRESS
 import com.karhoo.uisdk.util.TestData.Companion.SELECTED_ADDRESS
-import com.karhoo.uisdk.util.TestData.Companion.SHORT
 import com.karhoo.uisdk.util.TestData.Companion.THIRD_FLEET
 import com.karhoo.uisdk.util.TestData.Companion.TIMEOUT
 import com.karhoo.uisdk.util.TestData.Companion.TRIP
@@ -111,7 +108,7 @@ class AdyenBookingTests : Launch {
             reverseGeocodeResponse(HTTP_BAD_REQUEST, GENERAL_ERROR)
         }
         booking(this, null) {
-            waitFor(MEDIUM)
+            mediumSleep()
         } result {
             checkSnackbarWithText(R.string.K0001)
         }
@@ -130,7 +127,7 @@ class AdyenBookingTests : Launch {
             quoteIdResponse(HTTP_BAD_REQUEST, NO_AVAILABILITY)
         }
         booking(this, CLEAN_TRIP_INTENT) {
-            waitFor(SHORT)
+            shortSleep()
         } result {
             checkErrorIsShown(R.string.no_availability)
             contactButtonSnackbarIsEnabled()
@@ -145,7 +142,7 @@ class AdyenBookingTests : Launch {
             quoteIdResponse(HTTP_BAD_REQUEST, NO_AVAILABILITY)
         }
         booking(this, CLEAN_TRIP_INTENT) {
-            waitFor(MEDIUM)
+            mediumSleep()
             try {
                 //Send app to background
                 pressDeviceBackButton()
@@ -153,7 +150,7 @@ class AdyenBookingTests : Launch {
             }
         }
         booking(this, CLEAN_TRIP_INTENT) {
-            waitFor(SHORT)
+            shortSleep()
         } result {
             checkErrorIsShown(R.string.no_availability)
             contactButtonSnackbarIsEnabled()
@@ -193,7 +190,7 @@ class AdyenBookingTests : Launch {
             quotesResponse(HTTP_OK, VEHICLES_ASAP)
         }
         booking(this, INITIAL_TRIP_INTENT) {
-            waitFor(MEDIUM)
+            mediumSleep()
         } result {
             fullASAPQuotesListCheck()
         }
@@ -212,7 +209,7 @@ class AdyenBookingTests : Launch {
             quotesResponse(HTTP_OK, VEHICLES_ASAP)
         }
         booking(this, INITIAL_TRIP_INTENT) {
-            waitFor(SHORT)
+            shortSleep()
         } result {
             quotesListNotExpanded(THIRD_FLEET)
         }
@@ -236,7 +233,7 @@ class AdyenBookingTests : Launch {
             quotesResponse(HTTP_OK, VEHICLES_ASAP)
         }
         booking(this, INITIAL_TRIP_INTENT) {
-            waitFor(SHORT)
+            shortSleep()
             pressExpandListButton()
         } result {
             FLEET_INFO_ALT.name?.let { quotesListIsExpanded(it) }
@@ -261,7 +258,7 @@ class AdyenBookingTests : Launch {
         }
         booking(this, CLEAN_TRIP_INTENT) {
             pressPrebookButton()
-            waitFor(SHORT)
+            shortSleep()
         } result {
             localTimeMessageIsDisplayed()
         }
@@ -278,9 +275,9 @@ class AdyenBookingTests : Launch {
             reverseGeocodeResponse(HTTP_OK, REVERSE_GEO_SUCCESS)
         }
         booking(this, null) {
-            waitFor(SHORT)
+            shortSleep()
             pressPrebookButton()
-            waitFor(SHORT)
+            shortSleep()
             pressCancelPrebookWindow()
         } result {
             prebookWindowNotVisible()
@@ -298,11 +295,11 @@ class AdyenBookingTests : Launch {
             reverseGeocodeResponse(HTTP_OK, REVERSE_GEO_SUCCESS)
         }
         booking(this, null) {
-            waitFor(SHORT)
+            shortSleep()
             pressPrebookButton()
-            waitFor(SHORT)
+            shortSleep()
             pressOKPrebookWindow()
-            waitFor(SHORT)
+            shortSleep()
             pressCancelPrebookWindow()
         } result {
             prebookWindowNotVisible()
@@ -321,11 +318,11 @@ class AdyenBookingTests : Launch {
             reverseGeocodeResponse(HTTP_OK, REVERSE_GEO_SUCCESS)
         }
         booking(this, null) {
-            waitFor(SHORT)
+            shortSleep()
             pressPrebookButton()
-            waitFor(SHORT)
+            shortSleep()
             pressOKPrebookWindow()
-            waitFor(SHORT)
+            shortSleep()
             pressOKPrebookWindow()
         } result {
             prebookWindowNotVisible()
@@ -343,7 +340,7 @@ class AdyenBookingTests : Launch {
             reverseGeocodeResponse(HTTP_OK, REVERSE_GEO_SUCCESS)
         }
         booking(this, null) {
-            waitFor(SHORT)
+            shortSleep()
         } result {
             reverseGeoAddressVisiblePickUp(address = REVERSE_GEO_DISPLAY_ADDRESS)
         }
@@ -362,7 +359,7 @@ class AdyenBookingTests : Launch {
         }
         booking(this, CLEAN_TRIP_INTENT) {
             pressMenuButton()
-            waitFor(SHORT)
+            shortSleep()
             pressOutOfSideMenu()
         } result {
             sideMenuIsNotVisible()
@@ -378,7 +375,7 @@ class AdyenBookingTests : Launch {
     @AllowFlaky(attempts = 5)
     fun noPrebookButtonVisibleWhenNoAddressEntered() {
         booking(this, null) {
-            waitFor(SHORT)
+            shortSleep()
         } result {
             prebookButtonIsNotVisible()
         }
@@ -397,7 +394,7 @@ class AdyenBookingTests : Launch {
             quoteIdResponse(HTTP_BAD_REQUEST, ADDRESSES_IDENTICAL)
         }
         booking(this, BraintreeBookingTests.IDENTICAL_ADDRESSES_TRIP_INTENT) {
-            waitFor(MEDIUM)
+            shortSleep()
         } result {
             samePickUpAndDestinationErrorIsDisplayed()
         }
@@ -417,7 +414,7 @@ class AdyenBookingTests : Launch {
         booking(this) {
             pressPrebookButton()
             pressOKPrebookWindow()
-            waitFor(SHORT)
+            shortSleep()
             pressOKPrebookWindow()
         } result {
             prebookLogoNotVisible()
@@ -427,7 +424,7 @@ class AdyenBookingTests : Launch {
         }
         booking {
             clickOnLocateMeButton()
-            waitFor(SHORT)
+            shortSleep()
         } result {
             prebookLogoNotVisible()
         }
@@ -454,7 +451,7 @@ class AdyenBookingTests : Launch {
             prebookLogoNotVisible()
         }
         booking {
-            waitFor(MEDIUM)
+            mediumSleep()
             clickPickUpAddressField()
         }
         address {
@@ -462,7 +459,7 @@ class AdyenBookingTests : Launch {
             clickBakerStreetResult()
         }
         booking {
-            waitFor(SHORT)
+            shortSleep()
         } result {
             selectedPickupAddressIsVisible(address = SELECTED_ADDRESS)
             prebookLogoNotVisible()
@@ -482,9 +479,9 @@ class AdyenBookingTests : Launch {
             quotesResponse(HTTP_OK, VEHICLES_ASAP)
         }
         booking(this, INITIAL_TRIP_INTENT) {
-            waitFor(SHORT)
+            shortSleep()
             pressFirstQuote()
-            waitFor(MEDIUM)
+            mediumSleep()
         } result {
             bookARideScreenIsVisible()
         }
@@ -503,9 +500,9 @@ class AdyenBookingTests : Launch {
             quotesResponse(HTTP_OK, VEHICLES_ASAP)
         }
         booking(this, INITIAL_TRIP_INTENT) {
-            waitFor(SHORT)
+            shortSleep()
             pressFirstQuote()
-            waitFor(SHORT)
+            shortSleep()
             pressCloseBookARideScreen()
         } result {
             fullASAPQuotesListCheck()
@@ -529,9 +526,9 @@ class AdyenBookingTests : Launch {
             quotesResponse(HTTP_OK, VEHICLES_ASAP)
         }
         booking(this, INITIAL_TRIP_INTENT) {
-            waitFor(SHORT)
+            shortSleep()
             pressFirstQuote()
-            waitFor(MEDIUM)
+            shortSleep()
         } result {
             fullCheckBookARideScreenASAP()
         }
@@ -556,11 +553,11 @@ class AdyenBookingTests : Launch {
             bookingDetailsResponse(code = HTTP_OK, response = TRIP_DER_NO_NUMBER_PLATE, trip = TRIP.tripId)
         }
         booking(this, INITIAL_TRIP_INTENT) {
-            waitFor(SHORT)
+            shortSleep()
             pressFirstQuote()
-            waitFor(MEDIUM)
+            mediumSleep()
             pressBookRideButton()
-            waitFor(LONG)
+            longSleep()
         } result {
             checkDriverDetails()
         }
@@ -585,7 +582,7 @@ class AdyenBookingTests : Launch {
             clickOnProfileButton()
         }
         userProfile {
-            waitFor(MEDIUM)
+            mediumSleep()
         } result {
             fullScreenCheckCardRegisteredAdyen()
         }
