+++ uisdk/src/androidTest/java/com/karhoo/uisdk/trip/TripTests.kt
@@ -1,12 +1,13 @@
-package com.karhoo.uisdk.trip.adyenTrip
+package com.karhoo.uisdk.trip
 
 import android.content.Intent
 import android.os.Bundle
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import androidx.test.filters.LargeTest
 import androidx.test.rule.ActivityTestRule
 import androidx.test.rule.GrantPermissionRule
 import com.github.tomakehurst.wiremock.common.ConsoleNotifier
-import com.github.tomakehurst.wiremock.core.WireMockConfiguration
+import com.github.tomakehurst.wiremock.core.WireMockConfiguration.options
 import com.github.tomakehurst.wiremock.junit.WireMockRule
 import com.karhoo.sdk.api.model.TripInfo
 import com.karhoo.uisdk.R
@@ -15,22 +16,21 @@ import com.karhoo.uisdk.common.Launch
 import com.karhoo.uisdk.common.serverRobot
 import com.karhoo.uisdk.common.testrunner.UiSDKTestConfig
 import com.karhoo.uisdk.screen.trip.TripActivity
-import com.karhoo.uisdk.trip.trip
 import com.karhoo.uisdk.util.TestData
-import com.karhoo.uisdk.util.TestData.Companion.ADYEN_PROVIDER
-import com.karhoo.uisdk.util.TestData.Companion.ADYEN_PUBLIC_KEY
+import com.karhoo.uisdk.util.TestData.Companion.BRAINTREE_PROVIDER
+import com.karhoo.uisdk.util.TestData.Companion.BRAINTREE_TOKEN
 import com.karhoo.uisdk.util.TestData.Companion.DRIVER_TRACKING
 import com.karhoo.uisdk.util.TestData.Companion.GENERAL_ERROR
-import com.karhoo.uisdk.util.TestData.Companion.MEDIUM
 import com.karhoo.uisdk.util.TestData.Companion.QUOTE_LIST_ID_ASAP
 import com.karhoo.uisdk.util.TestData.Companion.REVERSE_GEO_SUCCESS
-import com.karhoo.uisdk.util.TestData.Companion.SHORT
 import com.karhoo.uisdk.util.TestData.Companion.TRIP
 import com.karhoo.uisdk.util.TestData.Companion.TRIP_ARRIVED
 import com.karhoo.uisdk.util.TestData.Companion.TRIP_CANCELLED_BY_FLEET
 import com.karhoo.uisdk.util.TestData.Companion.TRIP_DER
 import com.karhoo.uisdk.util.TestData.Companion.TRIP_DER_NO_DRIVER_DETAILS
 import com.karhoo.uisdk.util.TestData.Companion.TRIP_DER_NO_NUMBER_PLATE
+import com.karhoo.uisdk.util.TestData.Companion.TRIP_DER_NO_VEHICLE_AND_DRIVER_DETAILS
+import com.karhoo.uisdk.util.TestData.Companion.TRIP_DER_NO_VEHICLE_DETAILS
 import com.karhoo.uisdk.util.TestData.Companion.TRIP_DER_NO_VEHICLE_NUMBER_PLATE_AND_DRIVER_DETAILS
 import com.karhoo.uisdk.util.TestData.Companion.TRIP_POB
 import com.karhoo.uisdk.util.TestData.Companion.TRIP_STATUS_ARRIVED
@@ -53,7 +53,9 @@ import java.net.HttpURLConnection.HTTP_INTERNAL_ERROR
 import java.net.HttpURLConnection.HTTP_OK
 
 @RunWith(AndroidJUnit4::class)
-class AdyenTripTests : Launch {
+@LargeTest
+class TripTests : Launch {
+
     @get:Rule
     val activityRule: ActivityTestRule<TripActivity> =
             ActivityTestRule(TripActivity::class.java, false, false)
@@ -65,7 +67,7 @@ class AdyenTripTests : Launch {
             .around(activityRule)
 
     @get:Rule
-    var wireMockRule = WireMockRule(WireMockConfiguration.options()
+    var wireMockRule = WireMockRule(options()
                                             .port(UiSDKTestConfig.PORT_NUMBER)
                                             .notifier(ConsoleNotifier(true)))
 
@@ -77,8 +79,8 @@ class AdyenTripTests : Launch {
     fun setUp() {
         serverRobot {
             successfulToken()
-            paymentsProviderResponse(HTTP_OK, ADYEN_PROVIDER)
-            sdkInitResponse(HTTP_OK, ADYEN_PUBLIC_KEY)
+            paymentsProviderResponse(HTTP_OK, BRAINTREE_PROVIDER)
+            sdkInitResponse(HTTP_OK, BRAINTREE_TOKEN)
         }
     }
 
@@ -100,40 +102,11 @@ class AdyenTripTests : Launch {
                 tracking = GENERAL_ERROR,
                 details = GENERAL_ERROR)
         trip(this) {
-            sleep()
+            shortSleep()
         } result {
             checkSnackbarWithText(R.string.K0001)
         }
     }
-
-    /**
-     * Given:   The driver is en route
-     * When:    the fleet cancels the ride
-     * Then:    I see a message that the booking is not cancelled for cause of no drivers, alternative option is available
-     **/
-    @Test
-    @AllowFlaky(attempts = 10)
-    fun quoteAlternativeAfterFleetCancelledWhenDER() {
-        mockTripSuccessResponse(
-                status = TRIP_STATUS_CANCELLED_BY_FLEET,
-                tracking = DRIVER_TRACKING,
-                details = TRIP_CANCELLED_BY_FLEET)
-        serverRobot {
-            reverseGeocodeResponse(HTTP_OK, REVERSE_GEO_SUCCESS)
-            quoteIdResponse(HTTP_CREATED, QUOTE_LIST_ID_ASAP)
-            quotesResponse(HTTP_OK, VEHICLES_ASAP)
-        }
-        trip(this) {
-            waitFor(SHORT)
-            clickAlternativeButton()
-        }
-        booking {
-            waitFor(MEDIUM)
-        } result {
-            fullASAPQuotesListCheck()
-        }
-    }
-
 
     /**
      * Given:   A driver has been allocated
@@ -150,7 +123,7 @@ class AdyenTripTests : Launch {
                 details = TRIP_DER)
         trip(this) {
             clickOnDriverDetails()
-            waitFor(SHORT)
+            shortSleep()
         } result {
             driverDERDetailsFullCheck()
         }
@@ -190,7 +163,7 @@ class AdyenTripTests : Launch {
                 tracking = DRIVER_TRACKING,
                 details = TRIP_ARRIVED)
         trip(this) {
-            sleep()
+            shortSleep()
             clickOnDriverDetails()
         } result {
             driverArrivedDetailsFullCheck()
@@ -210,7 +183,7 @@ class AdyenTripTests : Launch {
                 tracking = DRIVER_TRACKING,
                 details = TRIP_ARRIVED)
         trip(this) {
-            sleep()
+            shortSleep()
         } result {
             ArrivedFullScreenCheck(
                     pickupText = TRIP_ARRIVED.origin?.displayAddress.orEmpty(),
@@ -251,11 +224,11 @@ class AdyenTripTests : Launch {
                 tracking = DRIVER_TRACKING,
                 details = TRIP_POB)
         trip(this) {
-            sleep()
+            shortSleep()
         } result {
             fullScreenCheckPOB(
-                    pickupText = TRIP_ARRIVED.origin?.displayAddress.orEmpty(),
-                    destinationText = TRIP_ARRIVED.destination?.displayAddress.orEmpty()
+                    pickupText = TestData.TRIP_ARRIVED.origin?.displayAddress.orEmpty(),
+                    destinationText = TestData.TRIP_ARRIVED.destination?.displayAddress.orEmpty()
                               )
         }
     }
@@ -274,7 +247,7 @@ class AdyenTripTests : Launch {
                 details = TRIP_ARRIVED)
         trip(this) {
             clickOnDriverDetails()
-            waitFor(MEDIUM)
+            mediumSleep()
             clickOnDriverDetails()
         } result {
             driverDetailsNoLongerExpanded()
@@ -372,7 +345,7 @@ class AdyenTripTests : Launch {
             clickOnDriverDetails()
             clickOnCancelRide()
             clickOnCancelYourRideCancellation()
-            sleep()
+            shortSleep()
         } result {
             checkRideInProgress()
         }
@@ -398,6 +371,38 @@ class AdyenTripTests : Launch {
     }
 
     /**
+     * Given:   The driver is en route
+     * When:    the fleet cancels the ride
+     * Then:    I see a message that the booking is not cancelled for cause of no drivers, alternative option is available
+     **/
+    @Test
+    @Ignore
+    @AllowFlaky(attempts = 5)
+    fun alternativeAfterFleetCancelledWhenDER() {
+        mockTripSuccessResponse(
+                status = TRIP_STATUS_CANCELLED_BY_FLEET,
+                tracking = DRIVER_TRACKING,
+                details = TRIP_CANCELLED_BY_FLEET)
+        serverRobot {
+            successfulToken()
+            sdkInitResponse(HTTP_OK, BRAINTREE_TOKEN)
+            paymentsProviderResponse(HTTP_OK, BRAINTREE_PROVIDER)
+            reverseGeocodeResponse(HTTP_OK, REVERSE_GEO_SUCCESS)
+            quoteIdResponse(HTTP_CREATED, QUOTE_LIST_ID_ASAP)
+            quotesResponse(HTTP_OK, VEHICLES_ASAP)
+        }
+        trip(this) {
+            shortSleep()
+            clickAlternativeButton()
+        }
+        booking {
+            mediumSleep()
+        } result {
+            fullASAPQuotesListCheck()
+        }
+    }
+
+    /**
      * Given:   The driver's name and details are unavailable
      * When:    I am on the trip screen and status is DER
      * Then:    I can still see the trip details box
@@ -413,8 +418,8 @@ class AdyenTripTests : Launch {
             clickOnDriverDetails()
         } result {
             DERFullScreenCheck(
-                    pickupText = TestData.TRIP_DER.origin?.displayAddress.orEmpty(),
-                    destinationText = TestData.TRIP_DER.destination?.displayAddress.orEmpty()
+                    pickupText = TRIP_DER.origin?.displayAddress.orEmpty(),
+                    destinationText = TRIP_DER.destination?.displayAddress.orEmpty()
                               )
             noDriverDetailsDERCheck()
         }
@@ -429,15 +434,15 @@ class AdyenTripTests : Launch {
     @AllowFlaky(attempts = 5)
     fun tripDetailsBoxVisibleWithoutVehicleDetails() {
         mockTripSuccessResponse(
-                status = TestData.TRIP_STATUS_DER,
-                tracking = TestData.DRIVER_TRACKING,
-                details = TestData.TRIP_DER_NO_VEHICLE_DETAILS)
+                status = TRIP_STATUS_DER,
+                tracking = DRIVER_TRACKING,
+                details = TRIP_DER_NO_VEHICLE_DETAILS)
         trip(this) {
             clickOnDriverDetails()
         } result {
             DERFullScreenCheck(
-                    pickupText = TestData.TRIP_DER.origin?.displayAddress.orEmpty(),
-                    destinationText = TestData.TRIP_DER.destination?.displayAddress.orEmpty()
+                    pickupText = TRIP_DER.origin?.displayAddress.orEmpty(),
+                    destinationText = TRIP_DER.destination?.displayAddress.orEmpty()
                               )
             noVehicleDetailsDERCheck()
         }
@@ -452,15 +457,15 @@ class AdyenTripTests : Launch {
     @AllowFlaky(attempts = 3)
     fun tripDetailsBoxVisibleWithoutVehicleAndDriverDetails() {
         mockTripSuccessResponse(
-                status = TestData.TRIP_STATUS_DER,
-                tracking = TestData.DRIVER_TRACKING,
-                details = TestData.TRIP_DER_NO_VEHICLE_AND_DRIVER_DETAILS)
+                status = TRIP_STATUS_DER,
+                tracking = DRIVER_TRACKING,
+                details = TRIP_DER_NO_VEHICLE_AND_DRIVER_DETAILS)
         trip(this) {
             clickOnDriverDetails()
         } result {
             DERFullScreenCheck(
-                    pickupText = TestData.TRIP_DER.origin?.displayAddress.orEmpty(),
-                    destinationText = TestData.TRIP_DER.destination?.displayAddress.orEmpty()
+                    pickupText = TRIP_DER.origin?.displayAddress.orEmpty(),
+                    destinationText = TRIP_DER.destination?.displayAddress.orEmpty()
                               )
             noVehicleAndDriverDetailsDERCheck()
         }
